Feature: Display Thanks for Flagging Message
Because:
* This gives the user visual feedback that their flag was successful and explains what happens next.

This Commit:
* Moves the flag modal to the container so only one of them is ever rendered.
* Stops sending flag notification to discord on development.
