+++ app/javascript/components/project-submissions/components/index.js
@@ -1,9 +1,11 @@
 import SubmissionsList from './submissions-list';
 import Modal from './modal';
 import CreateForm from './create-form';
+import FlagForm from './flag-form';
 
 export {
   SubmissionsList,
   Modal,
   CreateForm,
+  FlagForm,
 };
